refactor: extract `cacvote_server::client` into its own crate
