+++ apps/cacvote-jx-terminal/backend/Cargo.toml
@@ -15,7 +15,7 @@ auth-rs = { workspace = true }
 axum = { workspace = true }
 base64 = { workspace = true }
 base64-serde = { workspace = true }
-cacvote-server = { workspace = true }
+cacvote-server-client = { workspace = true }
 clap = { workspace = true }
 color-eyre = { workspace = true }
 dotenvy = { workspace = true }
