+++ apps/cacvote-server/backend/Cargo.toml
@@ -31,6 +31,7 @@ auth-rs = { workspace = true }
 axum = { workspace = true }
 base64 = { workspace = true }
 base64-serde = { workspace = true }
+cacvote-server-client = { workspace = true }
 clap = { workspace = true }
 color-eyre = { workspace = true }
 dotenvy = { workspace = true }
