+++ libs/cacvote-server-client/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "cacvote-server-client"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+reqwest = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+thiserror = { workspace = true }
+types-rs = { workspace = true }
+url = { workspace = true }
+uuid = { workspace = true }
